arm64: dts: qcom: sdm845-xiaomi-beryllium: Fix compatible strings
The most specific strings should come first (panel variant included),
then the more generic ones (device and SoC).

This is necessary for hexagonrpcd to guess the HexagonFS path for the
device. It tries all of the compatible entries, but if none for
xiaomi,beryllium existed it wouldn't be able to guess it.

Signed-off-by: Jens Reidel <adrian@travitia.xyz>